+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `/advise` now notifies when a reconsideration request is queued.
+
 ## [0.1.4] - 2026-08-30
 
 ### Fixed